Refactoring & E2E testing & Bugfixing

+++ internal/service/department_service.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"hierarchy-management/internal/domain"
+	"hierarchy-management/internal/errors"
 )
 
 type DepartmentService interface {
@@ -21,14 +22,30 @@ func NewDepartmentService(repo domain.DepartmentRepository) DepartmentService {
 }
 
 func (s *departmentService) CreateDepartment(dept *domain.Department) error {
+	if s.repo.IsDuplicateDepartmentName(dept.Name) {
+		return errors.NewDuplicateEntryError("Department", "name", dept.Name)
+	}
+
+	if dept.ParentID != nil && !s.repo.ExistsByID(*dept.ParentID) {
+		return errors.NewValidationError("Department", "Parent department does not exist")
+	}
+
 	return s.repo.Create(dept)
 }
 
 func (s *departmentService) UpdateDepartment(dept *domain.Department) error {
+	if !s.repo.ExistsByID(dept.ID) {
+		return errors.NewValidationError("Department", "Department does not exist")
+	}
+
 	return s.repo.Update(dept)
 }
 
 func (s *departmentService) DeleteDepartment(id int) error {
+	if !s.repo.ExistsByID(id) {
+		return errors.NewValidationError("Department", "Department does not exist")
+	}
+
 	return s.repo.Delete(id)
 }
